Safari-611.1.10/LayoutTests/editing/execCommand/indent-nested-inlines-1.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div contenteditable="true"><a href="#">Link</a>This should be under same the blockquote as Link</div>
<script>

Markup.description('This tests indenting paragraphs that begin with inlines.');

var div = document.getElementsByTagName('div')[0];
window.getSelection().selectAllChildren(div);
document.execCommand('indent', false, true);
document.execCommand('indent', false, true);
document.execCommand('indent', false, true);

Markup.dump(div);

</script>
</body>
</html>
